mac-capture: Fix display capture cursor option
The "show cursor" option does not work, and the display stream is
defaulting to its internal default, which on 10.8 and 10.9 is visible,
and 10.10+ is invisible.  The issue is that you cannot use @true or
@false with the dictionary being passed to the display stream.  After
some testing, it needs to point to the id of either kCFBooleanTrue or
kCFBooleanFalse in order for it to work properly.